Fix file validation, reorganize directories, update README.md
