+++ src/main/java/org/simplelibrary/config/SecurityConfiguration.java
@@ -1,5 +1,6 @@
-package org.simplelibrary.security;
+package org.simplelibrary.config;
 
+import org.simplelibrary.service.AccountDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
