+++ src/main/resources/application-prod.properties
@@ -10,7 +10,7 @@ spring.datasource.password=${sm://db_password}
 
 spring.resources.cache.cachecontrol.max-age=12h
 
-spring.servlet.multipart.max-file-size=5MB
-spring.servlet.multipart.max-request-size=5MB
+spring.servlet.multipart.max-file-size=50MB
+spring.servlet.multipart.max-request-size=50MB
 
 upload.path=gs://simple-library-media
